Widen the offset column, which only one of the two stores could refuse
The offsets in the committed session are around 6.46 billion. SQLite's INTEGER is eight bytes,
so every offline test passed on them; PostgreSQL's integer is four, and it rejected the first
row with 'integer out of range'.

This is the defect the second store leg exists to catch, found by it on its first run: a schema
that is correct in the store you develop against and wrong in the one you deploy to. Both tables
now say bigint, which both stores accept, and a test asserts both the column type and that the
committed session still carries offsets too large for four bytes, so it cannot pass by the data
getting smaller.

+++ src/queuez/consumer.py
@@ -30,12 +30,17 @@ SCHEMA_SQL = """
 create table if not exists bar (
     topic  text not null,
     domain text not null,
-    events integer not null,
+    events bigint not null,
     primary key (topic, domain)
 );
 create table if not exists applied (
     topic       text not null,
-    offset_seen integer not null,
+    -- BIGINT, AND `integer` WAS WRONG IN A WAY ONLY ONE STORE NOTICED. These offsets are around
+    -- 6.46 billion, which does not fit in PostgreSQL's four-byte integer. SQLite's INTEGER is
+    -- eight bytes, so the offline suite passed on every one of them and the server rejected the
+    -- first with `integer out of range`. A schema that is correct in the store you develop
+    -- against and wrong in the one you deploy to is the reason this leg exists.
+    offset_seen bigint not null,
     fingerprint text not null,
     primary key (topic, offset_seen)
 );
